Move locale-related functions from wagtail.admin.utils to wagtail.admin.locale
